requirements.txt
flask
gunicorn
openai
pinecone

numpy==1.26.4

langchain==0.1.16
langchain-community==0.0.32
langchain-openai==0.0.8
langchain-huggingface
sentence-transformers
python-dotenv
tiktoken
pypdf